make bucket modifiable

+++ 10_realtime/predict.sh
@@ -1,11 +1,18 @@
 #!/bin/bash
 
-gsutil -m rm -rf gs://cloud-training-demos-ml/flights/chapter10/output
+if [ "$#" -ne 1 ]; then
+    echo "Usage: ./predict.sh bucket-name"
+    exit
+fi
+
+BUCKET=$1
+
+gsutil -m rm -rf gs://$BUCKET/flights/chapter10/output
 bq rm -f flights.predictions
 
 cd chapter10
 
 mvn compile exec:java \
  -Dexec.mainClass=com.google.cloud.training.flights.AddRealtimePrediction \
- -Dexec.args="--realtime --speedupFactor=60 --maxNumWorkers=10 --autoscalingAlgorithm=THROUGHPUT_BASED"
+ -Dexec.args="--realtime --speedupFactor=60 --maxNumWorkers=10 --autoscalingAlgorithm=THROUGHPUT_BASED --bucket=$BUCKET"
